requirements.txt
aiohttp==3.8.4
aiosignal==1.3.1
altair==4.2.2
async-timeout==4.0.2
attrs==22.2.0
blinker==1.6.1
cachetools==5.3.0
certifi==2022.12.7
charset-normalizer==3.1.0
click==8.1.3
dataclasses-json==0.5.7
decorator==5.1.1
entrypoints==0.4
faiss-cpu==1.7.3
filelock==3.11.0
frozenlist==1.3.3
gitdb==4.0.10
GitPython==3.1.31
huggingface-hub==0.13.4
idna==3.4
importlib-metadata==6.2.1
Jinja2==3.1.2
joblib==1.2.0
jsonschema==4.17.3
langchain==0.0.135
markdown-it-py==2.2.0
MarkupSafe==2.1.2
marshmallow==3.19.0
marshmallow-enum==1.5.1
mdurl==0.1.2
mpmath==1.3.0
multidict==6.0.4
mypy-extensions==1.0.0
networkx==3.1
nltk==3.8.1
numpy==1.24.2
openai==0.27.4
openapi-schema-pydantic==1.2.4
packaging==23.0
pandas==1.5.3
Pillow==9.5.0
protobuf==3.20.3
pyarrow==11.0.0
pydantic==1.10.7
pydeck==0.8.0
Pygments==2.14.0
Pympler==1.0.1
pyrsistent==0.19.3
python-dateutil==2.8.2
pytz==2023.3
pytz-deprecation-shim==0.1.0.post0
PyYAML==6.0
regex==2023.3.23
requests==2.28.2
rich==13.3.3
scikit-learn==1.2.2
scipy==1.10.1
sentence-transformers==2.2.2
sentencepiece==0.1.97
six==1.16.0
smmap==5.0.0
SQLAlchemy==1.4.47
streamlit==1.21.0
sympy==1.11.1
tenacity==8.2.2
threadpoolctl==3.1.0
tokenizers==0.13.3
toml==0.10.2
toolz==0.12.0
torch==2.0.0
torchvision==0.15.1
tornado==6.2
tqdm==4.65.0
transformers==4.27.4
typing-inspect==0.8.0
typing_extensions==4.5.0
tzdata==2023.3
tzlocal==4.3
urllib3==1.26.15
validators==0.20.0
xmltodict==0.13.0
yarl==1.8.2
zipp==3.15.0
